requirements.txt
streamlit
google-generativeai
openai
firebase-admin
requests
streamlit-folium
folium
gtts
polyline
python-dotenv